[ESPEasySerial] Update to 2.0.10
