+++ lib/ESPEasySerial/library.json
@@ -1,6 +1,6 @@
 {
     "name": "ESPeasySerial",
-    "version": "2.0.9",
+    "version": "2.0.10",
     "keywords": [
       "serial", "io", "softwareserial", "hardwareserial"
     ],
