+++ lib/ESPEasySerial/library.properties
@@ -1,5 +1,5 @@
 name=ESPeasySerial
-version=2.0.9
+version=2.0.10
 author=Gijs Noorlander
 maintainer=Gijs Noorlander <gijs.noorlander@gmail.com>
 sentence=Wrapper for SoftwareSerial, SC16IS752 I2C to UART bridge and HardwareSerial for ESP8266 and ESP32/ESP32S2.
