+++ platformio_esp32_envs.ini
@@ -34,7 +34,7 @@ lib_deps                  =
                             adafruit/Adafruit BusIO @ ^1.10.0
                             adafruit/Adafruit GFX Library@^1.11.1
                             adafruit/Adafruit ILI9341@^1.5.10
-                            td-er/ESPeasySerial @ 2.0.9
+                            td-er/ESPeasySerial @ 2.0.10
                             td-er/SparkFun MAX1704x Fuel Gauge Arduino Library @ ^1.0.1
 extra_scripts             = post:tools/pio/post_esp32.py
                             ${extra_scripts_default.extra_scripts}
