+++ platformio_esp82xx_base.ini
@@ -68,7 +68,7 @@ lib_deps                  =
                             adafruit/Adafruit GFX Library@^1.11.1
                             adafruit/Adafruit ILI9341@^1.5.10
                             bblanchon/ArduinoJson @ ^6.17.2
-                            td-er/ESPeasySerial @ 2.0.9
+                            td-er/ESPeasySerial @ 2.0.10
                             td-er/RABurton ESP8266 Mutex @ ^1.0.2
                             td-er/SparkFun MAX1704x Fuel Gauge Arduino Library @ ^1.0.1
 lib_ignore                = ${esp82xx_defaults.lib_ignore}
